requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.21
reportlab==4.0.7
python-dateutil==2.8.2
Werkzeug==2.3.7
PyMySQL==1.1.2
gunicorn==21.2.0
python-dotenv==1.0.0
Pillow==10.0.0
